+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Change log
 
+## 0.3.11 27-Oct-2021
+
+- fix orderBy chain-able but no real effect
+- preventing user from chain `offset` after `offset`(reduce mistake, however cannot guarantee there is only 1 offset till the end of chain).
+- preventing user from chain `limit` or `limit to last` after `limit` or `limit to last`(reduce mistake, however cannot guarantee there is only 1 limit or limit to last till the end of chain).
+- fix "Type instantiation is excessively deep and possibly infinite", cause is known but found no explanation for this.
+- fix recursive type become `any` after declaration is emitted, that causing the build file query type is not working(becoming un-chain-able, a lot of works done to fix this).
+
 ## 0.3.2 26-Oct-2021
 
 fix `CheckObjectHasDuplicateEndName` not exported bug